requirements.txt
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
streamlit==1.41.1
schedule==1.2.2
playwright==1.49.1
openai>=1.0.0
google-genai>=1.0.0
